+++ src/app/(with-side)/[type]/[id]/page.tsx
@@ -6,7 +6,7 @@ import { useSearchParams, useParams, useRouter } from 'next/navigation';
 import ComponentFrame from '@/components/common/componentFrame';
 import RecommendedContent from '@/components/common/recommendedContent';
 import DetailContent from '@/components/features/detail/content/DetailContent';
-import CommentSection from '@/components/features/detail/comment/CommentSection';
+import CommentSection from '@/components/features/detail/comment/comment-section-temp';
 import FetchingFailedCard from '@/components/common/fetching-failed-card';
 import PaginationBar from '@/components/common/pagination-bar.tsx/pagination-bar';
