+++ src/components/common/floating-writing-button.-temp.tsx
