+++ src/services/config/constants.ts
@@ -1,2 +1,1 @@
 export const SERVER_URL = process.env.NEXT_PUBLIC_SERVER_URL;
-export const JWT = process.env.NEXT_PUBLIC_ACCESS_JWT;
